+++ packages/plugin-undo-redo/web/src/types.js
@@ -1,0 +1,1 @@
+export const UNDO_REDO_TYPE = 'wix-rich-content-undo-redo';